requirements.txt
django
djangorestframework
django-crispy-forms
requests
django-tables2
jsonpickle
gunicorn
django-environ
pytest-pythonpath
pytest-cov
mixer
pytest
django-debug-toolbar
celery
django-redis